Add utility functions and remove extra logging
